fix(#1795): overflow-safe v5 cell bounds guards + Float32/varint parity + cell_value split
Consolidated lane closing #1795, #1884, #1885.

length into usize and then evaluated `offset + total_len > data.len()` as a
bounds guard — the ADD overflowed usize on an adversarial length prefix (panic
`attempt to add with overflow`, found by the #1614 fuzz crate) before the guard
could reject. New shared helper `read_vint_length_prefixed_bytes` caps the raw
length against MAX_CELL_VALUE_LENGTH BEFORE the `as usize` cast (mirroring
parse_vint_length) and compares with saturating subtraction. Applied to the
blob/text/decimal/duration/counter arms and the frozen-UDT length reads.

(f32 as f64) — matching the block path.

blob default → Value::Blob).

Flipped the decoder_lockstep float/varint pins from recorded-divergence to
converged. Regression test pins adversarial lengths → Err (no panic).

Campsite split (#1116): cell_value.rs 1251 -> 341 lines; scalar arms ->
cell_value_scalar.rs (587), complex ladder -> cell_value_complex.rs (333).

Drive-by unblocks (pre-existing on origin/main, blocked core clippy/tests):
full_index_stream_tests.rs missing `use serial_test::serial` (#2437);
issue_1578 test .err().expect() -> .expect_err() (#2070).
