+++ cqlite-core/src/storage/sstable/reader/parsing/row_decoder/mod.rs
@@ -854,6 +854,9 @@ mod block_emit;
 mod block_emit_windowed;
 mod cell_kind;
 mod cell_value;
+// campsite split of `cell_value` (issue #1795): scalar arms + complex ladder.
+mod cell_value_complex;
+mod cell_value_scalar;
 mod compaction;
 mod compaction_stream; // issue #2299 (split of `compaction`, campsite #1116)
 pub(in crate::storage::sstable::reader) use compaction_stream::{
@@ -915,6 +918,11 @@ mod regression_1741k_tests;
 // allocating semantics (`!marker && parse_partition_header_full.is_ok()`).
 #[cfg(test)]
 mod regression_1641_boundary_peek_tests;
+
+// Issue #1795: per-cell VInt-length bounds guards reject adversarial lengths
+// (return `Err`, never overflow-panic).
+#[cfg(test)]
+mod regression_1795_overflow_tests;
 
 impl V5CompressedLegacyParser {
     /// Create a new V5CompressedLegacy parser
